requirements.txt
wxPython>=4.2.1
pyobjc-framework-Cocoa>=11.0; sys_platform == "darwin"
pyobjc-framework-AVFoundation>=11.0; sys_platform == "darwin"
